+++ libs/angular-components/community/components/form/file-dropzone/file-dropzone.component.ts
@@ -6,13 +6,16 @@ import {
   input,
   signal,
   viewChild,
-  model,
   ElementRef,
-  effect,
   OnInit,
   output,
+  forwardRef,
 } from "@angular/core";
-import { ControlValueAccessor, ReactiveFormsModule } from "@angular/forms";
+import {
+  ControlValueAccessor,
+  NG_VALUE_ACCESSOR,
+  ReactiveFormsModule,
+} from "@angular/forms";
 import {
   ClosingButtonComponent,
   FeedbackTextComponent,
@@ -31,6 +34,7 @@ import {
   FeedbackTextProps,
   FileDropzone,
   FileInputMode,
+  SizeDisplayStandard,
 } from "./types";
 import {
   formatBytes,
@@ -58,7 +62,14 @@ import { FileService } from "./file.service";
     VerticalSpacingDirective,
     TediTranslationPipe,
   ],
-  providers: [FileService],
+  providers: [
+    FileService,
+    {
+      provide: NG_VALUE_ACCESSOR,
+      useExisting: forwardRef(() => FileDropzoneComponent),
+      multi: true,
+    },
+  ],
 })
 export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
   /**
@@ -75,7 +86,21 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
    **/
   maxSize = input<number>(0);
   /**
+   * Specifies the standard to use when displaying file sizes or maximum file size.
+   * Options are "SI" (base 10) or "IEC" (base 2).
+   *
+   * SI units are in multiples of 1000 (e.g., 1 kB = 1000 bytes).
+   *
+   * IEC units are in multiples of 1024 (e.g., 1 KiB = 1024 bytes).
+   * @default "IEC"
+   *
+   * https://wiki.ubuntu.com/UnitsPolicy
+   */
+  sizeDisplayStandard = input<SizeDisplayStandard>("IEC");
+  /**
    * Determines if multiple files can be uploaded at once via the file picker.
+   *
+   * https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Attributes/multiple
    * @default false
    **/
   multiple = input<boolean>(false);
@@ -114,7 +139,6 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
    *  @default "append"
    */
   mode = input<FileInputMode>("append");
-
   /**
    * If true, allows uploading folders instead of just files.
    * This enables the user to select a folder and upload all its contents.
@@ -131,22 +155,13 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
     validateFileType,
   ]);
   /**
-   * Disables the file dropzone, preventing user interaction.
+   * If true, shows the file dropzone as in a erroring state with red border.
+   * Overrides default validation state.
    * @default false
-   */
-  disabled = model<boolean>(false);
+   **/
+  hasError = input<boolean>(false);
   /**
-   * Provides helper text or feedback (such as an error or instruction message) to guide the user.
-   */
-  uploadError = model<string>();
-  /**
-   * Provides helper text or feedback (such as an error or instruction message) to guide the user.
-   * By default this is automatically generated based on the `accept` and `maxSize` inputs.
-   */
-  helperText = model<FeedbackTextProps>();
-
-  /**
-   * Output event triggered when files are added or changed.
+   * Output event triggered when files are added.
    **/
   fileChange = output<FileDropzone[]>();
   /**
@@ -157,11 +172,10 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
   fileInputElement =
     viewChild.required<ElementRef<HTMLInputElement>>("fileInput");
 
-  formatBytes = (bytes: number): string => formatBytes(bytes);
+  formatBytes = (bytes: number): string =>
+    formatBytes(bytes, this.sizeDisplayStandard());
 
-  private _uploadState = computed(
-    () => this._fileService?.uploadState() || "none"
-  );
+  private _uploadState = computed(() => this._fileService.uploadState());
 
   private _onChange = (_: FileDropzone[]) => {};
   private _onTouched = () => {};
@@ -170,6 +184,10 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
   private _fileService = inject(FileService);
 
   isDragActive = signal<boolean>(false);
+  disabled = signal<boolean>(false);
+
+  uploadError = signal<string | null>(null);
+  files = this._fileService.files;
 
   classes = computed(() => {
     const classList = ["tedi-file-dropzone"];
@@ -177,7 +195,9 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
     if (this.disabled()) {
       classList.push("tedi-file-dropzone--disabled");
     }
-    if (this._uploadState() !== "none") {
+    if (this.hasError()) {
+      classList.push("tedi-file-dropzone--invalid");
+    } else if (this._uploadState() !== "none") {
       classList.push(`tedi-file-dropzone--${this._uploadState()}`);
     }
     if (this.isDragActive()) {
@@ -190,56 +210,42 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
     return classList.join(" ");
   });
 
-  constructor() {
-    effect(() => {
-      this._fileService.maxSize = this.maxSize();
-      this._fileService.accept = this.accept();
-      this._fileService.mode = this.mode();
-      this._fileService.validators = this.validators();
-    });
+  helperText = computed<FeedbackTextProps>(() => ({
+    text: getDefaultHelpers(
+      this.accept(),
+      this.maxSize(),
+      this.sizeDisplayStandard(),
+      this._translationService.translate.bind(this._translationService)
+    ),
+    type: "hint",
+    position: "left",
+  }));
+
+  ngOnInit(): void {
+    this.addFiles(this.defaultFiles());
+    this._fileService.maxSize = this.maxSize;
+    this._fileService.accept = this.accept;
+    this._fileService.mode = this.mode;
+    this._fileService.validators = this.validators;
+    this._fileService.sizeDisplayStandard = this.sizeDisplayStandard;
   }
 
   fileClasses = (file: FileDropzone): string => {
     const classList = ["tedi-file-dropzone__file-item"];
-
     if (file.className) {
       classList.push(...file.className);
     }
-
     if (file.fileStatus != "none") {
       classList.push(`tedi-file-dropzone__file-item--${file.fileStatus}`);
     }
-
     return classList.join(" ");
   };
 
   tooltipClasses = (file: FileDropzone): string => {
     const classes = ["tedi-file-dropzone__tooltip"];
-    if (file.helper?.type) {
-      classes.push(
-        "tedi-file-dropzone__tooltip--" + file.helper.type || "hint"
-      );
-    }
-
+    classes.push(`tedi-file-dropzone__tooltip--${file.helper?.type || "hint"}`);
     return classes.join(" ");
   };
-
-  files = this._fileService.files;
-
-  ngOnInit(): void {
-    this.addFiles(this.defaultFiles());
-    if (!this.helperText()) {
-      this.helperText.set({
-        text: getDefaultHelpers(
-          this.accept(),
-          this.maxSize(),
-          this._translationService.translate.bind(this._translationService)
-        ),
-        type: "hint",
-        position: "left",
-      });
-    }
-  }
 
   selectionChange = (event: Event) => {
     const fileList = (event.target as HTMLInputElement).files;
@@ -300,7 +306,7 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
 
     this._updateErrorState(error);
 
-    this._onChange(normalizedFiles);
+    this._onChange(this._fileService.files());
     this.fileChange.emit(normalizedFiles);
   }
 
@@ -308,6 +314,7 @@ export class FileDropzoneComponent implements ControlValueAccessor, OnInit {
     const error = this._fileService.removeFiles([file]);
     this._updateErrorState(error);
 
+    this._onChange(this._fileService.files());
     this.fileDelete.emit(file);
   }