+++ libs/angular-components/community/components/form/file-dropzone/file.service.ts
@@ -3,32 +3,34 @@ import {
   DropzoneValidatorFunction,
   FileDropzone,
   FileInputMode,
+  SizeDisplayStandard,
   ValidationState,
 } from "./types";
 import { TediTranslationService } from "@tehik-ee/tedi-angular/tedi";
 
-@Injectable({
-  providedIn: "root",
-})
+@Injectable()
 export class FileService {
-  maxSize = 0;
-  accept = "";
-  mode: FileInputMode = "append";
+  maxSize = signal(0).asReadonly();
+  accept = signal("").asReadonly();
+  mode = signal<FileInputMode>("append").asReadonly();
+  validators = signal<DropzoneValidatorFunction[]>([]).asReadonly();
+  sizeDisplayStandard = signal<SizeDisplayStandard>("IEC").asReadonly();
+
   uploadState = signal<ValidationState>("none");
-  validators: DropzoneValidatorFunction[] = [];
 
   protected _files = signal<FileDropzone[]>([]);
 
   private _translateService = inject(TediTranslationService);
 
   get files(): Signal<FileDropzone[]> {
-    return this._files;
+    return this._files.asReadonly();
   }
+
   public async addFiles(files: FileDropzone[] | File[]): Promise<string[]> {
     let newFiles = this.normalizeFiles(files);
     const currentFiles = this.files();
 
-    switch (this.mode) {
+    switch (this.mode()) {
       case "append": {
         // index any duplicate name file
         newFiles = await this._renameDuplicates(currentFiles, newFiles);
@@ -57,8 +59,8 @@ export class FileService {
       newFiles = newFiles.filter((file) => file.fileStatus !== "invalid");
     }
     const error = this._checkErrorState(newFiles);
-    this.uploadState.set(this._getNewState(!!error.length));
     this._files.set(newFiles);
+    this.uploadState.set(this._getNewState(!!error.length));
     return error;
   }
 
@@ -111,12 +113,13 @@ export class FileService {
     const errors: string[] = [];
     for (const file of files) {
       file.helper = undefined;
-      const error = this.validators
+      const error = this.validators()
         .map((validator) =>
           validator(
-            this.maxSize,
-            this.accept,
+            this.maxSize(),
+            this.accept(),
             file,
+            this.sizeDisplayStandard(),
             this._translateService.translate.bind(this._translateService)
           )
         )